Create Home component and use stack navigator in app
